Create LevelExtractionEvents
